add dryrun

+++ agentd/backup.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+import argparse
 import api
 import agentlib
 import datetime
@@ -13,7 +14,7 @@ logging.basicConfig(
 
 _logger = logging.getLogger(__name__)
 
-def create_scheduled_backups():
+def create_scheduled_backups(dryrun=False):
     today = datetime.date.today()
     if today.day == 1:
         if (today.month-1) % 3 == 0:
@@ -29,13 +30,18 @@ def create_scheduled_backups():
     _logger.info(f"Creating scheduled backups for {len(insts)} instances.")
     for instance in insts:
         uid = instance['uid']
-        api.backup(uid, trigger=trigger)
+        if not dryrun:
+            api.backup(uid, trigger=trigger)
         backups_paths = glob.glob(agentlib.dump_path(uid, trigger, '*.pgc'))
         backups_paths.sort(reverse=True) # reverse=True -> desc -> largest/latest first
         n = 3 # Remove all except n latest dumps.
         for remove in backups_paths[n:]:
             _logger.info("Remove %s", remove)
-            os.remove(remove)
+            if not dryrun:
+                os.remove(remove)
 
 if __name__ == "__main__":
-    create_scheduled_backups()
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--dryrun', '--dry-run', action='store_true')
+    args = parser.parse_args()
+    create_scheduled_backups(dryrun=args.dryrun)
